Edumacation scoring and move scoring into seperate file
