+++ src/Application/Features/Assessments/DTOs/V1/Pathways/Education/EducationPathway.cs
@@ -1,8 +1,6 @@
-using Cfo.Cats.Application.Features.Assessments.DTOs.V1.Pathways.Money;
-
 namespace Cfo.Cats.Application.Features.Assessments.DTOs.V1.Pathways.Education;
 
-public class EducationPathway : PathwayBase
+public partial class EducationPathway : PathwayBase
 {
     public EducationPathway()
     {
@@ -27,8 +25,5 @@ public class EducationPathway : PathwayBase
     public override string Title => "Education";
     public override double Constant => 0.61259;
     public override string Icon => CatsIcons.Education;
-    protected override IEnumerable<double> GetPercentiles(int age, AssessmentLocation location, Sex sex)
-    {
-        yield return 0;
-    }
+    
 }
